requirements.txt
openai
tqdm
python-dotenv
pandas
tiktoken
langchain
